backend/requirements.txt
Django
djangorestframework
PyJWT
requests
django-cors-headers
psycopg
gunicorn
django-environ
rest-client
djangorestframework-simplejwt
#dos2unix